requirements.txt
fastapi>=0.116.1
assemblyai>=0.42.1
google-generativeai>=0.8.5
httpx>=0.28.1
python-dotenv>=1.1.1
uvicorn>=0.35.0
python-multipart>=0.0.20
pydantic>=2.11.7
pydantic-settings>=2.0.0
